auth is finished

+++ frontend/src/components/WorkoutDetails.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 import { useNavigate } from 'react-router-dom';
 import { Link } from 'react-router-dom';
 
@@ -8,10 +9,20 @@ const WorkoutDetails = ({workout}) => {
 
 
   const {dispatch} = useWorkoutsContext() 
-  
+const {user}= useAuthContext()
+
+
   const handleClick = async() => {
+
+    if(!user){
+      return
+    }
+
    const response = await fetch('/api/workouts/'+workout._id, {
     method : 'DELETE',
+    headers:{
+      'Authorization': `Bearer ${user.token}`
+    }
    })  
    const json = await response.json()
